+++ native-port/divoomd/src/transport.rs
@@ -6,6 +6,7 @@ pub enum DeviceTransport {
     Ble(BleTransport),
     Spp(crate::spp::SppTransport),
     Lan(crate::lan::LanTransport),
+    Mock(crate::mock_transport::MockTransport),
 }
 
 impl DeviceTransport {
@@ -15,6 +16,7 @@ impl DeviceTransport {
             DeviceTransport::Ble(b) => b.device_name(),
             DeviceTransport::Spp(s) => s.device_name(),
             DeviceTransport::Lan(_) => None,
+            DeviceTransport::Mock(m) => m.device_name(),
         }
     }
 
@@ -24,6 +26,7 @@ impl DeviceTransport {
             DeviceTransport::Ble(b) => b.set_cached_device_name(_name),
             DeviceTransport::Spp(s) => s.set_cached_device_name(_name),
             DeviceTransport::Lan(_) => {}
+            DeviceTransport::Mock(m) => m.set_cached_device_name(_name),
         }
     }
 
@@ -33,6 +36,7 @@ impl DeviceTransport {
             DeviceTransport::Ble(b) => b.send_command(command_id, args, write_with_response).await,
             DeviceTransport::Spp(s) => s.send_command(command_id, args, write_with_response).await.map_err(|e| e.to_string().into()),
             DeviceTransport::Lan(_) => Err("send_command not supported on LAN".into()),
+            DeviceTransport::Mock(m) => m.send_command(command_id, args, write_with_response).await,
         }
     }
 
@@ -42,6 +46,7 @@ impl DeviceTransport {
             DeviceTransport::Ble(b) => b.wait_for_response(command_id, timeout).await,
             DeviceTransport::Spp(s) => s.wait_for_response(command_id, timeout).await,
             DeviceTransport::Lan(_) => None,
+            DeviceTransport::Mock(m) => m.wait_for_response(command_id, timeout).await,
         }
     }
 
@@ -51,6 +56,7 @@ impl DeviceTransport {
             DeviceTransport::Ble(b) => b.send_command_and_wait(command_id, args, timeout).await,
             DeviceTransport::Spp(s) => s.send_command_and_wait(command_id, args, timeout).await,
             DeviceTransport::Lan(_) => None,
+            DeviceTransport::Mock(m) => m.send_command_and_wait(command_id, args, timeout).await,
         }
     }
 
@@ -60,6 +66,7 @@ impl DeviceTransport {
             DeviceTransport::Ble(b) => b.stream_animation_8b(blob).await,
             DeviceTransport::Spp(s) => s.stream_animation_8b(blob).await.map_err(|e| e.to_string().into()),
             DeviceTransport::Lan(_) => Err("stream_animation_8b not supported on LAN".into()),
+            DeviceTransport::Mock(m) => m.stream_animation_8b(blob).await,
         }
     }
 
@@ -69,6 +76,7 @@ impl DeviceTransport {
             DeviceTransport::Ble(_) => None,
             DeviceTransport::Spp(_) => None,
             DeviceTransport::Lan(l) => Some(l),
+            DeviceTransport::Mock(_) => None,
         }
     }
 
@@ -82,6 +90,7 @@ impl DeviceTransport {
             DeviceTransport::Ble(b) => b.wait_for_any_response(command_ids, timeout).await,
             DeviceTransport::Spp(s) => s.wait_for_any_response(command_ids, timeout).await,
             DeviceTransport::Lan(_) => None,
+            DeviceTransport::Mock(m) => m.wait_for_any_response(command_ids, timeout).await,
         }
     }
 }